backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
mysql-connector-python
python-multipart
passlib[bcrypt]
python-jose[cryptography]
Werkzeug
python-dotenv